Release 1.2
